requirements.txt
Flask==3.0.3
Flask_Bcrypt==1.0.1
Flask_Login==0.6.3
Flask_Mail==0.9.1
flask_sqlalchemy==3.1.1
flask_wtf==1.2.1
itsdangerous==2.2.0
tensorflow==2.15.0
tensorflow_estimator==2.15.0
numpy==1.26.4
opencv_python==4.9.0.80
python-dotenv==1.0.1
WTForms==3.1.2
psycopg2-binary==2.9.9
gunicorn==22.0.0

